requirements.txt
openai>=1.0.0
moviepy>=1.0.3
google-auth>=2.20.0
google-auth-oauthlib>=0.8.0
google-api-python-client>=2.70.0
requests>=2.30.0
Pillow>=9.5.0
python-dotenv>=1.0.0
tqdm>=4.65.0
